pyproject.toml
[project]
name = "cohechat"
version = "0.1.0"
description = ""
authors = [
    {name = "CheHyeonyeong"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "sqlmodel (>=0.0.27,<0.0.28)",
    "sqlalchemy-utc (>=0.14.0,<0.15.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "greenlet (>=3.2.4,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
